remove capnproto
